Driver bug fix
SPI driver fix and tuned Peak Detect param, now the FIFO data matches Max peak data

+++ ADuCM3029_Asset_Health_Core0/include/platform_drivers.h
@@ -140,7 +140,7 @@ int32_t spi_remove(spi_desc *desc);
 /* Write and read data to/from SPI. */
 int32_t spi_write_and_read(spi_desc *desc,
 			   uint8_t *data,
-			   uint8_t bytes_number);
+			   uint16_t bytes_number);
 
 /* Obtain the GPIO decriptor. */
 int32_t gpio_get(gpio_desc **desc,
